fix EIO_LIBS contents to proper variable.
vtorri: I told you these names are very very error prone. I did the
mistake while reading it before, raster did the same today and even
yourself did a typo. The "pc", "libs" and similar in the middle of a
long variable name is easy to miss! Even more if the prefix is
"requirements_" :-(



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@80234 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33