edje: make it possible to not reset body position on  state changes
To make it possible, a flag was added: ignore_part_position.
Enabled by default, it will let part position be update
only by physics calculatios. If the body is intended to be
moved when state is changed, the flag must to be explicitely
disabled.
It is required otherwise all the states would inherit position
from "default" and reset body's position.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@80276 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33