EPhysics: fix issues identified by edje support
Add slices as member of obj smart parent - when it's the case -
so we can restack it when the received evas object is member of a
smart object.

Resize the evas object on ephysics_body_evas_object_set to w=1,
h=1 when its initial/original size is 0, this prevents errors on
evas_object_image_source_set().

Reset the soft body transform on resize - put it back to its
prior position - the one before scaling.

On body_move do not consider the moving of anchored rigid body
enough, move the soft body vertices as well.

--This line, and those below, will be ignored--


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@80612 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33