efl: Introducing ecore_audio - the audio API for efl
This is still very much a work in progress, so expect some issues. The
signalling is using ecore events for now - that will change to callbacks
you can register callbacks for events on specific Ecore_Audio_Objects.
EO wasn't there when Ecore_Audio started, but it will probably move to
that in the future.

Otherwise have fun, don't break it (too much) and please send bug
reports and feedback to me.

Signed-off-by: Daniel Willmann <d.willmann@samsung.com>

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@80994 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33