efl/evas_object_image_is_inside: fixed implementation.
This function was basically never working correctly. Everything was
fixed by simulating the evas_object_image_render() workflow, but
instead of actually draw we just check the pixel transparency.

Bugs fixed:

 * fails when image is scaled up (could segv) or down (incorrect values);
 * fails when image is moved to negative x,y;
 * fails when border was being used.

Now everything is fixed and seems to work properly, except I'm not
handling the map and get_pixels() cases, these are marked with ERR()
so we can fix them if someone needs.




git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@81410 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33